squash: implement mounts for compose-based do; fix test_import_demo_course
