+++ tutor/commands/compose.py
@@ -9,7 +9,7 @@ from tutor import config as tutor_config
 from tutor import env as tutor_env
 from tutor import fmt, hooks, jobs, serialize, utils
 from tutor.commands.context import BaseJobContext
-from tutor.commands.do import do_command
+from tutor.commands.tasks import add_tasks_as_subcommands
 from tutor.exceptions import TutorError
 from tutor.types import Config
 
@@ -306,7 +306,7 @@ def importdemocourse(context: click.Context) -> None:
         f"""'tutor {dev_or_local} importdemocourse' has been renamed to 'tutor {dev_or_local} do importdemocourse'.
    'tutor {dev_or_local} importdemocourse' (without 'do') will stop working in a future release."""
     )
-    do_importdemocourse: click.Command = do_command.get_command(context, "importdemocourse")  # type: ignore
+    do_importdemocourse: click.Command = do.get_command(context, "importdemocourse")  # type: ignore
     context.invoke(do_importdemocourse)
 
 
@@ -332,6 +332,30 @@ def run(
     if not utils.is_a_tty():
         extra_args.append("-T")
     context.invoke(dc_command, command="run", args=[*extra_args, *args])
+
+
+@click.group(
+    help="Run a predefined task in new containers",
+    subcommand_metavar="TASKNAME [ARGS] ...",
+)
+@mount_option
+def do(mounts: t.Tuple[t.List[MountParam.MountType]]) -> None:
+    """
+    A command group for predefined tasks: `tutor (dev|local) do TASKNAME ARGS`
+    """
+    # Process mounts before handling any subcommand.
+    process_mount_arguments(mounts)
+
+
+@hooks.Actions.PLUGINS_LOADED.add()
+def _populate_do_after_plugins_loaded() -> None:
+    """
+    Dynamically populate the 'do' command group based on CLI_TASKS.
+
+    We do this after plugins are loaded to ensure that all plugins have had a chance
+    to add their entries to CLI_TASKS.
+    """
+    add_tasks_as_subcommands(do)
 
 
 @click.command(
@@ -561,7 +585,7 @@ def add_commands(command_group: click.Group) -> None:
     command_group.add_command(settheme)
     command_group.add_command(dc_command)
     command_group.add_command(run)
-    command_group.add_command(do_command)
+    command_group.add_command(do)
     command_group.add_command(copyfrom)
     command_group.add_command(bindmount_command)
     command_group.add_command(execute)
